Moves some tests from PrepareContractionHierarchiesTest to NodeContractorTest and cleans up.
* these tests basically 'misuse' PCH only to set the node priorities to the
maximum level

* makes PCH#initFromGraph,prepareNodes,contractNodes private

* removes PCH#testFindShortcuts

* some clean-up
