+++ core/src/test/java/com/graphhopper/routing/ch/NodeContractorTest.java
@@ -17,18 +17,24 @@
  */
 package com.graphhopper.routing.ch;
 
+import com.graphhopper.routing.Dijkstra;
+import com.graphhopper.routing.DijkstraOneToMany;
 import com.graphhopper.routing.util.CarFlagEncoder;
 import com.graphhopper.routing.util.EncodingManager;
 import com.graphhopper.routing.util.TraversalMode;
 import com.graphhopper.routing.weighting.ShortestWeighting;
 import com.graphhopper.routing.weighting.Weighting;
 import com.graphhopper.storage.*;
+import com.graphhopper.util.CHEdgeIteratorState;
+import com.graphhopper.util.EdgeIteratorState;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.*;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertTrue;
 
 public class NodeContractorTest {
     private final CarFlagEncoder encoder = new CarFlagEncoder();
@@ -36,13 +42,183 @@ public class NodeContractorTest {
     private final Weighting weighting = new ShortestWeighting(encoder);
     private final GraphHopperStorage g = new GraphBuilder(encodingManager).setCHGraph(weighting).create();
     private final CHGraph lg = g.getGraph(CHGraph.class);
-    private final TraversalMode traversalMode = TraversalMode.NODE_BASED;
+    private final TraversalMode tMode = TraversalMode.NODE_BASED;
     private Directory dir;
 
     @Before
     public void setUp() {
         dir = new GHDirectory("", DAType.RAM_INT);
     }
+
+    private void createExampleGraph() {
+        //5-1-----2
+        //   \ __/|
+        //    0   |
+        //   /    |
+        //  4-----3
+        //
+        g.edge(0, 1, 1, true);
+        g.edge(0, 2, 1, true);
+        g.edge(0, 4, 3, true);
+        g.edge(1, 2, 3, true);
+        g.edge(2, 3, 1, true);
+        g.edge(4, 3, 2, true);
+        g.edge(5, 1, 2, true);
+        g.freeze();
+    }
+
+    @Test
+    public void testShortestPathSkipNode() {
+        createExampleGraph();
+        double normalDist = new Dijkstra(g, weighting, tMode).calcPath(4, 2).getDistance();
+        DijkstraOneToMany algo = new DijkstraOneToMany(g, weighting, tMode);
+        CHGraph lg = g.getGraph(CHGraph.class);
+
+        setMaxLevelOnAllNodes();
+
+        algo.setEdgeFilter(new NodeContractor.IgnoreNodeFilter(lg, g.getNodes() + 1).setAvoidNode(3));
+        algo.setWeightLimit(100);
+        int nodeEntry = algo.findEndNode(4, 2);
+        assertTrue(algo.getWeight(nodeEntry) > normalDist);
+
+        algo.clear();
+        algo.setMaxVisitedNodes(1);
+        nodeEntry = algo.findEndNode(4, 2);
+        assertEquals(-1, nodeEntry);
+    }
+
+    @Test
+    public void testShortestPathSkipNode2() {
+        createExampleGraph();
+        double normalDist = new Dijkstra(g, weighting, tMode).calcPath(4, 2).getDistance();
+        assertEquals(3, normalDist, 1e-5);
+        DijkstraOneToMany algo = new DijkstraOneToMany(g, weighting, tMode);
+
+        setMaxLevelOnAllNodes();
+
+        algo.setEdgeFilter(new NodeContractor.IgnoreNodeFilter(lg, g.getNodes() + 1).setAvoidNode(3));
+        algo.setWeightLimit(10);
+        int nodeEntry = algo.findEndNode(4, 2);
+        assertEquals(4, algo.getWeight(nodeEntry), 1e-5);
+
+        nodeEntry = algo.findEndNode(4, 1);
+        assertEquals(4, algo.getWeight(nodeEntry), 1e-5);
+    }
+
+    @Test
+    public void testShortestPathLimit() {
+        createExampleGraph();
+        DijkstraOneToMany algo = new DijkstraOneToMany(g, weighting, tMode);
+
+        setMaxLevelOnAllNodes();
+
+        algo.setEdgeFilter(new NodeContractor.IgnoreNodeFilter(lg, g.getNodes() + 1).setAvoidNode(0));
+        algo.setWeightLimit(2);
+        int endNode = algo.findEndNode(4, 1);
+        // did not reach endNode
+        assertNotEquals(1, endNode);
+    }
+
+    @Test
+    public void testDirectedGraph() {
+        //5 6 7
+        // \|/
+        //4-3_1<-\ 10
+        //     \_|/
+        //   0___2_11
+
+        g.edge(0, 2, 2, true);
+        g.edge(10, 2, 2, true);
+        g.edge(11, 2, 2, true);
+        // create a longer one directional edge => no longish one-dir shortcut should be created
+        g.edge(2, 1, 2, true);
+        g.edge(2, 1, 10, false);
+
+        g.edge(1, 3, 2, true);
+        g.edge(3, 4, 2, true);
+        g.edge(3, 5, 2, true);
+        g.edge(3, 6, 2, true);
+        g.edge(3, 7, 2, true);
+        g.freeze();
+
+        setMaxLevelOnAllNodes();
+
+        // find all shortcuts if we contract node 1
+        NodeContractor nodeContractor = new NodeContractor(dir, g, lg, weighting, tMode);
+        nodeContractor.initFromGraph();
+        Collection<NodeContractor.Shortcut> scs = nodeContractor.testFindShortcuts(1);
+        assertEquals(2, scs.size());
+        Iterator<NodeContractor.Shortcut> iter = scs.iterator();
+        NodeContractor.Shortcut sc1 = iter.next();
+        NodeContractor.Shortcut sc2 = iter.next();
+        if (sc1.weight > sc2.weight) {
+            NodeContractor.Shortcut tmp = sc1;
+            sc1 = sc2;
+            sc2 = tmp;
+        }
+
+        // both dirs
+        assertTrue(sc1.toString(), sc1.from == 3 && sc1.to == 2);
+        assertTrue(sc1.toString(), encoder.isForward(sc1.flags) && encoder.isBackward(sc1.flags));
+
+        // directed
+        assertTrue(sc2.toString(), sc2.from == 2 && sc2.to == 3);
+        assertTrue(sc2.toString(), encoder.isForward(sc2.flags));
+
+        assertEquals(sc1.toString(), 4, sc1.weight, 1e-4);
+        assertEquals(sc2.toString(), 12, sc2.weight, 1e-4);
+    }
+
+    @Test
+    public void testFindShortcuts_Roundabout() {
+        EdgeIteratorState iter1_3 = g.edge(1, 3, 1, true);
+        EdgeIteratorState iter3_4 = g.edge(3, 4, 1, true);
+        EdgeIteratorState iter4_5 = g.edge(4, 5, 1, false);
+        EdgeIteratorState iter5_6 = g.edge(5, 6, 1, false);
+        EdgeIteratorState iter6_8 = g.edge(6, 8, 2, false);
+        EdgeIteratorState iter8_4 = g.edge(8, 4, 1, false);
+        g.edge(6, 7, 1, true);
+        g.freeze();
+
+        CHEdgeIteratorState tmp = lg.shortcut(1, 4);
+        tmp.setFlags(PrepareEncoder.getScDirMask());
+        tmp.setWeight(2);
+        tmp.setSkippedEdges(iter1_3.getEdge(), iter3_4.getEdge());
+        long f = PrepareEncoder.getScFwdDir();
+        tmp = lg.shortcut(4, 6);
+        tmp.setFlags(f);
+        tmp.setWeight(2);
+        tmp.setSkippedEdges(iter4_5.getEdge(), iter5_6.getEdge());
+        tmp = lg.shortcut(6, 4);
+        tmp.setFlags(f);
+        tmp.setWeight(3);
+        tmp.setSkippedEdges(iter6_8.getEdge(), iter8_4.getEdge());
+
+        setMaxLevelOnAllNodes();
+
+        lg.setLevel(3, 3);
+        lg.setLevel(5, 5);
+        lg.setLevel(7, 7);
+        lg.setLevel(8, 8);
+
+        // there should be two different shortcuts for both directions!
+        NodeContractor nodeContractor = new NodeContractor(dir, g, lg, weighting, tMode);
+        nodeContractor.initFromGraph();
+        Collection<NodeContractor.Shortcut> sc = nodeContractor.testFindShortcuts(4);
+        assertEquals(2, sc.size());
+        Iterator<NodeContractor.Shortcut> iter = sc.iterator();
+        NodeContractor.Shortcut sc1 = iter.next();
+        NodeContractor.Shortcut sc2 = iter.next();
+        if (sc1.from > sc2.from) {
+            NodeContractor.Shortcut tmpSc = sc1;
+            sc1 = sc2;
+            sc2 = tmpSc;
+        }
+
+        assertEquals("1->6, weight:4.0 (7,8)", sc1.toString());
+        assertEquals("6->1, weight:5.0 (9,7)", sc2.toString());
+    }
+
 
     @Test
     public void testShortcutMergeBug() {
@@ -64,8 +240,15 @@ public class NodeContractorTest {
         NodeContractor.Shortcut sc2 = new NodeContractor.Shortcut(1, 3, 6.82048125, 121.25);
         sc2.flags = PrepareEncoder.getScDirMask();
         List<NodeContractor.Shortcut> list = Arrays.asList(sc1, sc2);
-        NodeContractor nodeContractor = new NodeContractor(dir, g, lg, weighting, traversalMode);
+        NodeContractor nodeContractor = new NodeContractor(dir, g, lg, weighting, tMode);
         nodeContractor.initFromGraph();
         assertEquals(2, nodeContractor.addShortcuts(list));
+    }
+
+    private void setMaxLevelOnAllNodes() {
+        int nodes = lg.getNodes();
+        for (int node = 0; node < nodes; node++) {
+            lg.setLevel(node, nodes + 1);
+        }
     }
 }
